▣ Contextualization logic inclusion.
● Key points:

• Now, the EntityMetadata and other classes related metadata work as Dto. When the code will treat the metadata, the original is converted to ContextualizationMetadata, that is updated between the flow of query contruction to fill the necessary requests.
