+++ src/services/mod.rs
@@ -4,8 +4,9 @@ pub mod query {
     }
 
     pub mod database {
+
         pub mod sqlite {
-            pub mod sqlite_query;
+            pub mod query;
 
             pub mod common {
                 pub mod tokens {
@@ -25,8 +26,20 @@ pub mod query {
             }
 
             pub mod operations {
+
+                pub mod expand {
+                    pub mod expand;
+
+                    pub mod token {
+                        pub mod tokenization;
+                    }
+                    pub mod context {
+                        pub mod context;
+                    }
+                }
+
                 pub mod select {
-                    pub mod select_sqlite;
+                    pub mod select;
 
                     pub mod token {
                         pub mod tokenization;
@@ -34,23 +47,15 @@ pub mod query {
                 }
     
                 pub mod filter {
-                    pub mod filter_sqlite;
+                    pub mod filter;
     
                     pub mod token {
                         pub mod tokenization;
                     }
                 }
-    
-                pub mod expand {
-                    pub mod expand_sqlite;
-
-                    pub mod token {
-                        pub mod tokenization;
-                    }
-                }
-    
+   
                 pub mod orderby {
-                    pub mod orderby_sqlite;
+                    pub mod orderby;
 
                     pub mod token {
                         pub mod tokenization;
@@ -58,41 +63,80 @@ pub mod query {
                 }
     
                 pub mod top {
-                    pub mod top_sqlite;
+                    pub mod top;
                 }
     
                 pub mod skip {
-                    pub mod skip_sqlite;
+                    pub mod skip;
                 }
             } 
         }
     
         pub mod mssql {
-            pub mod mssql_query;
+            pub mod query;
 
-            pub mod select {
-                pub mod select_mssql;
+            pub mod common {
+                pub mod tokens {
+                    pub mod filter {
+                        pub mod token;
+                    }
+                    pub mod expand {
+                        pub mod token;
+                    }
+                    pub mod select {
+                        pub mod token;
+                    }
+                    pub mod orderby {
+                        pub mod token;
+                    }
+                }
             }
 
-            pub mod filter {
-                pub mod filter_mssql;
-            }
+            pub mod operations {
 
-            pub mod expand {
-                pub mod expand_mssql;
-            }
+                pub mod expand {
+                    pub mod expand;
 
-            pub mod orderby {
-                pub mod orderby_mssql;
-            }
+                    pub mod token {
+                        pub mod tokenization;
+                    }
+                    pub mod context {
+                        pub mod context;
+                    }
+                }
 
-            pub mod top {
-                pub mod top_mssql;
-            }
+                pub mod select {
+                    pub mod select;
 
-            pub mod skip {
-                pub mod skip_mssql;
-            }
+                    pub mod token {
+                        pub mod tokenization;
+                    }
+                }
+    
+                pub mod filter {
+                    pub mod filter;
+    
+                    pub mod token {
+                        pub mod tokenization;
+                    }
+                }
+   
+                pub mod orderby {
+                    pub mod orderby;
+
+                    pub mod token {
+                        pub mod tokenization;
+                    }
+                }
+    
+                pub mod top {
+                    pub mod top;
+                }
+    
+                pub mod skip {
+                    pub mod skip;
+                }
+            } 
         }
     } 
 }
