--- src/services/query/database/mssql/top/top_mssql.rs
@@ -1,28 +1,0 @@
-use crate::dto::metadata::EntityMetadata;
-
-pub struct TopMssql;
-
-impl TopMssql {
-
-    pub fn validate<T: EntityMetadata>(
-        top: &i32
-    ) -> Result<(), String> {
-        
-        if top.to_owned() < 0 {
-            return Err("Invalid $top: must be a non-negative integer".into());
-        }
-
-        Ok(())
-    }
-    
-    pub fn process<T: EntityMetadata>(
-        top: &i32
-    ) -> Result<String, String> {
-        
-        let mut sql: String = String::new();
-
-        sql.push_str(&format!("\nLIMIT {}", top));
-    
-        Ok(sql)
-    }
-}
